+++ src/base/widget.c
@@ -3052,20 +3052,7 @@ ret_t widget_set_style(widget_t* widget, const char* state_and_name, const value
 
   widget_invalidate(widget, NULL);
 
-  if (strstr(name, "_color") != NULL && value->type == VALUE_TYPE_STRING) {
-    value_t v;
-    color_t c = color_parse(value_str(value));
-    value_set_uint32(&v, c.color);
-
-    return style_set(widget->astyle, state, name, &v);
-  } else if (tk_str_eq(name, STYLE_ID_BORDER) && value->type == VALUE_TYPE_STRING) {
-    value_t v;
-    value_set_uint32(&v, border_from_str(value_str(value)));
-
-    return style_set(widget->astyle, state, name, &v);
-  } else {
-    return style_set(widget->astyle, state, name, value);
-  }
+  return style_set(widget->astyle, state, name, value);
 }
 
 ret_t widget_set_style_int(widget_t* widget, const char* state_and_name, int32_t value) {