+++ tests/style_mutable_test.cc
@@ -100,11 +100,11 @@ TEST(StyleMutable, copy) {
   style_t* m2 = style_mutable_create(NULL, NULL);
 
   style_mutable_set_int(m1, "normal", "font_size", 123);
-  style_mutable_set_str(m1, "normal", "font", "foo");
+  style_mutable_set_str(m1, "normal", "font_name", "foo");
   style_mutable_set_color(m1, "normal", "text_color", c1);
 
   style_mutable_set_int(m1, "focused", "font_size", 321);
-  style_mutable_set_str(m1, "focused", "font", "foo2");
+  style_mutable_set_str(m1, "focused", "font_name", "foo2");
   style_mutable_set_color(m1, "focused", "text_color", c2);
   style_mutable_copy(m2, m1);
 
@@ -114,9 +114,10 @@ TEST(StyleMutable, copy) {
   style_mutable_foreach(m2, on_style_item, &str2);
 
   ASSERT_EQ(str1, str2);
-  ASSERT_EQ(str1,
-            string("normal,font_size,123;normal,font,\"foo\";normal,text_color,67305985;focused,"
-                   "font_size,321;focused,font,\"foo2\";focused,text_color,67305986;"));
+  ASSERT_EQ(
+      str1,
+      string("normal,font_size,123;normal,font_name,\"foo\";normal,text_color,67305985;focused,"
+             "font_size,321;focused,font_name,\"foo2\";focused,text_color,67305986;"));
 
   style_destroy(m1);
   style_destroy(m2);